Added fetcher module for retrieving remote files via HTTP
--HG--
rename : views/__init__.py => __init__.py

+++ models/tests/test_mainmodel.py
@@ -55,21 +55,22 @@ class TestMainModel(unittest.TestCase):
             for root, dirs, files in os.walk(diconde_folder):
                 for fname in files:
                     dicom_data_file = os.path.join(root, fname)
-                    #dicom_data_file = os.path.join(os.path.dirname(__file__), 'support_files',
-                    #    'dicondeDxExampleImage000.dcm')
-                    dicom_data = dicom.read_file(dicom_data_file)
-                    dicom_arr = dicom_data.pixel_array
-                    try:
-                        self.model.import_dicom(dicom_data_file)
-                    except TypeError:
-                        print(dicom_data_file)
-                    dest_file = os.path.join(pathfinder.data_path(),
-                        os.path.basename(dicom_data_file))
-                    self.assertTrue(os.path.exists(dest_file))
-                    read_data = np.loadtxt(dest_file, delimiter=',')
-                    self.assertListEqual(dicom_arr.tolist(), read_data.tolist())
-                    if os.path.exists(dest_file):
-                        os.remove(dest_file)
+                    basename, ext = os.path.splitext(dicom_data_file)
+                    # Simple check to ensure we're looking at DICOM files
+                    if ext.lower() == '.dcm':
+                        dicom_data = dicom.read_file(dicom_data_file)
+                        dicom_arr = dicom_data.pixel_array
+                        try:
+                            self.model.import_dicom(dicom_data_file)
+                        except TypeError:
+                            print(dicom_data_file)
+                        dest_file = os.path.join(pathfinder.data_path(),
+                            os.path.basename(dicom_data_file))
+                        self.assertTrue(os.path.exists(dest_file))
+                        read_data = np.loadtxt(dest_file, delimiter=',')
+                        self.assertListEqual(dicom_arr.tolist(), read_data.tolist())
+                        if os.path.exists(dest_file):
+                            os.remove(dest_file)
         except ImportError:
             return
